requirements.txt
pandas
numpy
kaggle
torch 
torchvision
ipykernel
pretrainedmodels
matplotlib
seaborn
opencv-python
Pillow